Don't push artifacts to AWS/Azure unless on main/master
In #1962 and #1963, we recently updated the isMainBranch variable to treat `release/*` branches as "main". This was so that we run the "full" integration tests suite on merges to release branches.

However, this change meant we also started pushing those assets to Azure and AWS, meaning the assets there are a mix of both `release/2.0` bits and `master` bits. This makes some of the other automation problematic, so we should revert to only pushing those bits if we're on `master`.