[2.0] Handle multiple tracer instances gracefully (#1990)
* Add TracerManager for managing the lifetime of "inner" tracer objects

The Tracer instance now delegates most of its behaviour to the TracerManager, and instead primarily becomes a "helper" on top of IScopeManager.

This changes the "main" API for users from being the Tracer constructor, to the static Tracer.ReplaceGlobalSettings(). This makes the real behaviour clear. We maintain the existing constructor for compatibility, but mark it obsolete.

The TracerManager is responsible for cleaning up resources when we replace global agentwriter/sampler etc. Currently it will _always_ replace the services, though we could make this smarter and not replace them if nothing has changed (which would reduce some overhead on startup for users that just setting a couple of things in code, like the service name)

Note that we now create a singleton IScopeManager, as logs injection with lib log can not support multiple concurrent traces.

Had to refactor Api.cs so as to not rely on Tracer.Instance.

This commit temporarily removes the LibLog integration, it will be added back in the next commit

* Update LibLogScopeEventSubscriber implementation

It doesn't actually require a tracer instance, it just requires a scope manager and the Tracing Without limits properties. This correctly creates a new lib log instance if settings change, and disposes the old one.

* Update tests and test apps

- Add a helper method that avoids the old constructors (which replace the global tracer)
- Replace some usages of the old APIs
- Replace usages of Tracer.Instance that cause issues in CI otherwise
- Fix snapshots

* Update SpanExtensions to explicitly pass a tracer instance

In the diagnostic observer (for example) this previously resulted in creating a separate Tracer instance (in practical terms, only during tests, but better to be explicit I think)

* #pragma ignore uses of obsolete Tracer.Instance set in benchmarks

* Move cleanup to a separate Task as it requires async work

* Fix cleanup of old TracerManager to be proper fire and forget

* Replace usages of obsolete API in benchmarks

* Pass TracerSettings into SpanExtensions.SetHttpCode instead of Tracer

* Use a callback to update sample rates instead of an ISampler instance

* Fix sampling rate test

* Feedback from PR