Add ImmutableTracerSettings
Changing TracerSettings after they've been passed to a Tracer instance is not valid and could lead to unpredictable results. To try and emphasise that, this adds a new ImmutableTracerSettings type, that does not allow changing any properties

For backwards compatability, I've kept the TracerSettings object as not-obsolete, as the de-facto "builder" for ImmutableTracerSettings. This also avoids the need to use any base classes, avoids having to duplicate initialization code, and also avoids the need for a dedicated "builder" type for the immutable settings (we can't use `init set;` properties because of missing references unfortunately)