Replace usages of IntegrationInfo with IntegrationIds
Thanks to the version conflict resolution, we will never have integrations where the integration name is unknown. Therefore we don't need to keep the dictionary of names, and we can use IntegrationIds directly instead of IntegrationInfo

If a customer is using manual tracing, and calls TracerSettings.Integrations["SomeIntegration"] with an unknown integration name, we will now just return a "default" IntegrationSettings, and log a warning.

Replaced all the old usages of IntegrationInfo with IntegrationIds