[2.0] Add .NET 6 WebApplicationBuilder tests (#2016)
* Add .NET 6 minimal APIs sample project

* Add integration tests for .NET 6 minimal APIs

Note that the NoFF /api/delay/0 tests have a different resource name than the equivalent ApiController version:  `GET /api/delay/?` instead of `GET api/delay/{seconds}`. This is because we don't run the endpoint routing events, so we end up using the "default" resource names, which don't include the route templates, even though they're technically available.

* Add snapshots for .net 6 webapplication/minimal api tests

* Add explicit DiagnosticObserver tests for .NET 6's WebApplicationBuilder

The most annoying thing here is how WebApplication adds an implicit UseRouting call at the start of the pipeline (before our existing middleware). That messes up the `UseMultipleErrorHandlerPipelines()` helper we had for testing multiple types of pipeline re-execution.

The upshot is that to handle the same paths, we need to build the pipeline in slightly different ways. This is "fine", just a bit annoying. I couldn't see a good alternative way around it, other than separating the executions to entirely different apps (but didn't want to that refactor here)

* Don't try and run diff tool inside CI