Remove blocking (#2011)
The blocking was removed because we made a poor choice when deciding to instrument Microsoft.AspNetCore.Builder.ApplicationBuilder.Build as this method is called several times on startup. Without this instrumentation blocking will not work on AspNetCore so it mades sense to remove it everywhere.

The current WAF definition never blocks, so having this functionality is not currently useful.