[2.0] Various small DiagnosticObserver optimisations (#2024)
* Make EndpointFeatureProxy a virtual class instead of an interface as slightly faster

* Use request.PathBase.Value not request.PathBase

Otherwise we fall back to string.Format()

* Use separate ASP.NET Core tags implementations

These more correctly represent the tags we can (and want) to set on the inner tags

* Calculate "obfuscated" urls on request end, not request start

We always throw away the calculation _except_ in error (404) cases or where middleware returns the response. By calculating it on exit, we avoid wasted work.

I'm not entirely clear of the Azure Functions behaviour here, so kept existing behaviour for that.

Also minor refactor to move the integrationsEnabled check outside of AspNetCoreHttpRequestHandler so we don't check it twice

* Remove unneeded properties from RequestTrackingFeature

We record these on the tags of the initial span anyway, so no need to duplicate

* Avoid calling HttpContext.Request.GetUrl() where possible

As this allocates. Do our best to avoid allocation by using PathBase and Path directly, recording the original values in the tracking feature.

This assumes that the Scheme and Host don't change, but those seem like reasonable assumptions to make here

* Remove unused field

* Use TryDuckCast instead of DuckCast

* Fix HttpRequestIn_PopulateSpan test as we no longer populate resource name at the start of the pipeline