[2.0] Add ImmutableTracerSettings (#1951)
* Add ImmutableTracerSettings

Changing TracerSettings after they've been passed to a Tracer instance is not valid and could lead to unpredictable results. To try and emphasise that, this adds a new ImmutableTracerSettings type, that does not allow changing any properties

For backwards compatability, I've kept the TracerSettings object as not-obsolete, as the de-facto "builder" for ImmutableTracerSettings. This also avoids the need to use any base classes, avoids having to duplicate initialization code, and also avoids the need for a dedicated "builder" type for the immutable settings (we can't use `init set;` properties because of missing references unfortunately)

* Make some mutable properties immutable

Update the readonly properties on ImmutableTracerSettings to be "immutable" collections. The biggest change here is ImmutableTracerSettingsCollection, which is a little bit clumsy, but seemed to be the best way to make Integrations immutable-ish.

* Add test that immutable tracer settings is immutable

* Update public API

* Add unit test to confirm ImmutableTracerSettings contains all (most) of the properties on TracerSettings

* ImmutableIntegrationSettings is actually immutable!

* Delete AdoNetExcludedTypes as will be removed in #2008

* Use IReadOnlyDictionary for HeaderTags and GlobalTags

* Revert "Delete AdoNetExcludedTypes as will be removed in #2008"

This reverts commit 8c6602d1c19c55084c66a5153fe437ae1fd22257.

* Fix public API