Remove the newline between the log message and properties
Currently the properties (e.g. `{ MachineName: ".", Process: "[44096 iisexpress]", AppDomain: "[2 /LM/W3SVC/1/ROOT-1-132820633973714685]", TracerVersion: "1.30.0.0" }`) are written on a new line _after_ the log line. This can make it difficult to tell whether a given set of properties is related to the line above or the line below. This change removes the newline.