Redesign package version generator to allow more granular control (#2091)
* Bump package versions with existing task

* Delete unused file

* Rename Program -> PackageVersionGenerator

* Expand PackageVersionEntry format

This allows us to be more flexible about which package versions we run. This is necessary as certain frameworks that we support cease being supported by NuGet packages (e.g. npgsl 6.0 isn't supported on netcoreapp2.1)

Takes the target frameworks from the samples themselves, so shouldn't need updating often

* Run GeneratePackageVersions with new format

The generated files are much more granular, but should be equivalent

* Downgrade tests on MSTest to 2.2.7 - there's a bug in the instrumentation of 2.2.8