Update supported ADO.NET versions (#2077)
* Bump supported versions of ADO.NET providers

* Update generator definitions

Note that we have not been testing the latest Npgsql release

Also Npgsql 6.0.0 does not support netcoreapp2.1 or netcoreapp3.0

* Update package versions

* Don't use Assembly.LoadFrom() with v4.0.0 as it crashes

Looks unrelated to the profiler, prob a bug in 4.0.0

* Version 4.0.0 doesn't work on Linux (can't connect)

* Explicitly set CheckEolTargetFramework=false in linux sample builds

For some reason these aren't picking it up from elsewhere

* Microsoft.Data.Sqlite hangs on Alpine with 6.0.0 on .NET Core 3.0

So skip the test. It hangs whether we are attached or not, so skip the test.