Don't update the dependabot file when updating integration definitions
Ideally, we should update this file whenever we update the versions we're _testing_ OR when we add a new integration. In that respect, updating the dependabot file when we run `UpdateIntegrationDefinitions` makes sense, but the downside is that this may suggest that we are _testing_ more than we really are.

> This also resulting in the dependabot file being changed in version-bump commits despite the fact we were neither changing integrations or testing more. This was essentially incorrect.

Instead, updating the dependabot as part of the `GeneratePackageVersions` target which updates what we test makes more sense to me.