Add log formatter
`LogFormatter` takes care of rendering logs with the correct name, and ensures we add all the additional TWoL tags as required.

The format is based on the Serilog RenderedCompactJsonFormatter, as this is known to be supported by the back-end.

The formatter provides utilities for writing properties and values to JSON. The EventIdHash provides a more compact view of the Message template to reduce the size of the logs.

LogPropertyRenderingDetails is necessary due to the fact that we may have already added env/service etc as part of the log event's properties, so we need to tell the formatter not to add them.

The RenderProperties function is a delegate so that we can use the `in` modifier, to reduce struct copies when the state is a readonly struct