Add sink implementation
The sink is responsible for queueing logs and sending batches using the ILogsApi (analogous the Agent.IApi type.

The batching code is single threaded, so we can reuse the same string builder throughout. Ideally this would all be serializing directly to utf8 bytes, which we _could_ consider ifdef-ing for .netcoreapp3.1.