Add initial NLog direct submission implementation
This was a real pain, as NLog's API changes in breaking ways even between minor versions.

The overall approach is to add a `Target` to the logging configuration every time a new logger is requested (typically once per instance)

Ultimately needed ~3 different versions for <4.3, 4.3-4.5, and 4.5+. (I stopped short of supporting 1.0.0.55, as that would require _another_ implementation).

There are a few hacky things here, in that NLog seems to expect Targets to be created once per logger, but we don't need to do that. As we also miss the typical point where targets are added, we have to call `Initialize()` on ourselves to ensure the target is activated correctly.

We also have to serialize the events immediately (instead of deferring to the background thread), as the events aren't properly otherwise (e.g. span_id and trace_id aren't captured etc)