Direct log submission (#2240)
* Add base batching API, based on Serilog.Sinks.PeriodicBatching

This is a heavily modified version of Serilog.Sinks.PeriodicBatching. It removes various scenarios we don't need to support, and replaces Serilog's BatchedConnectionStatus with an alternative CircuitBreaker implementation. Retries etc will be handled inside derived implementations instead.

* Add configuration for direct log submission

* Add log formatter

`LogFormatter` takes care of rendering logs with the correct name, and ensures we add all the additional TWoL tags as required.

The format is based on the Serilog RenderedCompactJsonFormatter, as this is known to be supported by the back-end.

The formatter provides utilities for writing properties and values to JSON. The EventIdHash provides a more compact view of the Message template to reduce the size of the logs.

LogPropertyRenderingDetails is necessary due to the fact that we may have already added env/service etc as part of the log event's properties, so we need to tell the formatter not to add them.

The RenderProperties function is a delegate so that we can use the `in` modifier, to reduce struct copies when the state is a readonly struct

* Add sink implementation

The sink is responsible for queueing logs and sending batches using the ILogsApi (analogous the Agent.IApi type.

The batching code is single threaded, so we can reuse the same string builder throughout. Ideally this would all be serializing directly to utf8 bytes, which we _could_ consider ifdef-ing for .netcoreapp3.1.

* Configure the global log sink in TelemetryManager

Currently direct log submission can only be done using external configuration, which simplifies things in a number of ways:
- We only have to initialize/final flush the logs API once
- Never have to worry about concurrent batch submission
- We can cache references to the sink and settings in all the appender/target integrations, as it won't change

If we make any of these settings configurable in code, or begin to rely on settings that _could_ change in code, then we may have to revisit this.

Note that users _can_ set the env/service/version in code, so we can't generally cache references to the `LogFormatter` (as it adds those values automatically)

* Add a MockLogsIntake for integration tests

* Add ILogger Direct submission implementation

This is slightly tricky because:
* ILoggerFactory uses a different "root" constructor in .net 5+
* We can't reverse duck type the ILoggerProvider directly, because we need to implement two interfaces, so have to `DefineType` an interface that implements both, and then duck type that.

* Add Serilog Direct submission implementation

This handles the case where they may have already the external "Serilog.Sinks.Datadog.Logs.DatadogSink" too, so they don't get duplicate logs

* Add log4net direct submission implementation

Instruments the AppenderCollection.ToArray() method to sneak our own appender in every time a logger is requested

* Add initial NLog direct submission implementation

This was a real pain, as NLog's API changes in breaking ways even between minor versions.

The overall approach is to add a `Target` to the logging configuration every time a new logger is requested (typically once per instance)

Ultimately needed ~3 different versions for <4.3, 4.3-4.5, and 4.5+. (I stopped short of supporting 1.0.0.55, as that would require _another_ implementation).

There are a few hacky things here, in that NLog seems to expect Targets to be created once per logger, but we don't need to do that. As we also miss the typical point where targets are added, we have to call `Initialize()` on ourselves to ensure the target is activated correctly.

We also have to serialize the events immediately (instead of deferring to the background thread), as the events aren't properly otherwise (e.g. span_id and trace_id aren't captured etc)

* Update integration definitions

* Apply suggestions from code review

Co-authored-by: Kevin Gosse <krix33@gmail.com>
Co-authored-by: Pierre Bonet <pierre.bonet@datadoghq.com>

* Update tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/Logging/Serilog/DirectSubmission/Formatting/SequenceValueDuck.cs

Co-authored-by: Pierre Bonet <pierre.bonet@datadoghq.com>

* Update tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/Logging/Serilog/DirectSubmission/Formatting/SerilogLogFormatter.cs

Co-authored-by: Kevin Gosse <krix33@gmail.com>

* Update to use throw helper where possible

* Add logFormatter tests

* Add named parameters for LogPropertyRenderingDetails ctor

* Rename `ddservice` to `service`

Not sure where I got `ddservice` from...

* Replace MethodGroup with Lambda

* Invert logic in BatchingSink.FlushBuffersTaskLoopAsync()

* Cleanup BoundedConcurrentQueue

- Remove Count so we're not tempted to use it. Expose inner queue for testing
- Remove try finally. These were added to be defensive against ThreadAbort, but as other places likely have far worse corruption from ThreadAbort, there's not much point in precious here. Worst case, `_counter` is off by 1, which likely isn't a big deal.

* Fix DatadogSink array growth logic

Previous logic didn't guarantee we would allocate an array large enough

* Make WriterWrapper a class instead of readonly struct

With the previous implementation, we were allocating a List<object>() for every struct, even if the list was never used (it's only used when there are non-dictionary scope values added using e.g. `ILogger.BeginScope("SomeValue")`)

With the new approach, we lazily allocate the list on first usage instead

* Record whether we're using the legacy appenders for NLog and log4Net

* Avoid infinite loop when `requiredTotalSize == MaxTotalSizeBytes`

* Remove allocation from DirectSubmissionLoggerProvider.CreateLogger()

* Load the API key in the normal way

* Fix incorrect type in logger

Co-authored-by: Kevin Gosse <krix33@gmail.com>
Co-authored-by: Pierre Bonet <pierre.bonet@datadoghq.com>