Update TelemetrySettings to use better URL inferring
If a DD_API key is provided, then assume we're running in an environment without an agent, and send directly to the intake.

If the user provides an override URL, use that.

If not, and the user provides DD_SITE, then use that for the intake. Avoids requiring manual config for telemetry e.g. in AAS (as DD_SITE) must always be configured. If DD_SITE is not provided, fall back to "datadoghq.com"

This logic matches that used in the Rust library.