Generate name of profiler jobs
Github Actions jobs are recognized by their name (or key if the name
is not present).
When there is a matrix (example: x64,x86 / net45,net5.0), there will
be as many jobs as number of combination of the matrix.
The final name will be "name (XX, YY)" (ex: "build (x64, net45)").