Add Telemetry (#2241)
* Allow filtering unit tests in Nuke

* Disable telemetry by default in integration tests

* Initial Telemetry implementation

* Add Samples.Telemetry and Telemetry integration tests (tests httpclient telemetry)

* Automatically set integration enabled/disabled for calltarget integrations

* Update with latest changes from Telemetry spec

- Endpoint has changed
- Data structures have changed
- Required headers have changed

* Allow using agent for sending telemetry

The agent will now (soon) accept telemetry data, so we should default to using the agent URL.

* If not using the agent intake, need to provide DD-API-KEY header in requests

* Send _latest_ settings to telemetry.

Important when the user is updating settings in code, now that we have a singleton Tracer at any one time.

* If there's a fatal error when sending telemetry, terminate the processing loop.

Currently I'm terminating the controller, as that removes the dependency event handler as well as stopping the processing loop. I think that is probably sufficient. We will continue to record when integrations start etc, but that info will never be used.

* Add manual recording of telemetry to initial integrations

This feels clumsy, but there's not really an "automatic" place to hook into, so this seems like the only valid approach.

Simplified the Curl/Sockets/Win handler implementations slightly which makes it easier to record the telemetry too.

* Add telemetry to integration tests for all instrumentation

* Try updating telemetry sample

* Disable telemetry for in-process instances

* Make ITelemetryController implement IDisposable

* Exclude "has error" from `Enabled` property condition

* Don't call `Abort()` on `HttpListener`

It throws a nullrefexception (sometimes) in net5.0+. Switch to using `Stop()` instead

* Handle errors when pushing telemetry

There are several cases where we _expect_ telemetry to fail to send. The most common will likely when the agent version isn't new enough, so we will get a 404 when sending.

To handle these "fatal" errors in addition to normal transient errors, we use the following approach.

- If we receive two 404s in a row from the endpoint, _without_ ever getting a success, we stop trying to push telemetry.
- If we receive five errors in a row from the endpoint, at any time, we stop trying to push telemetry.

If we have a transient error, we don't retry immediately. Instead, we hold onto the config/dependencies/integrations data, and use that to build the next request, unless these change in the mean-time. Currently, this is fixed at 1 minute between requests, as per the spec.

Note that for simplicity, we're currently still _recording_ telemetry, we're just not doing anything with it once we have too many failures. A subsequent improvement would be to trigger a replacement of the TelemetryController with the NullTelemetryController, but that's a little more complex to do.

* Improve efficiency of IntegrationTelemetryCollector

 Relies on the fact that in most cases, the integration will already have been executed.

 Based on micro benchmarks by @kevingosse (`Exchange` was original implementation, `Optimistic` is implementation in this commit)

 ```
 |      Method |      Mean |     Error |    StdDev |
 |------------ |----------:|----------:|----------:|
 |    Exchange | 4.3036 ns | 0.0500 ns | 0.0418 ns |
 |  Optimistic | 0.0513 ns | 0.0364 ns | 0.0284 ns |
 ```

* Record CallTarget telemetry in OnMethodEnd too

Necessary for instrumentations that don't use OnMethodBegin

* Clear previous config on FatalError

* Update tracer/src/Datadog.Trace/Telemetry/TelemetrySettings.cs

Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>

* Add telemetry testing for FakeDbCommand

* Update TelemetryCircuitBreakerTests.cs

* Add additional guard clauses and tests to TelemetryDataBuilder

* Rename telemetry Url -> Telemetry Uri

* Enable null references in TelemetryData DTOs

Requires a bunch of refactoring

* Update Telemetry config data to latest spec

Telemetry configuration data was recently changed to be an array of tuples, instead of an object. This updates the implementation and tests to handle that case.

* Apply suggestions from code review around integration id finding

Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>

* Add additional tests for TelemetryTransport failure paths

Convert socket exceptions (unable to connect) into fatal errors and handle 404 exceptions from WebRequest

* Update Telemetry loop to not push until initialization is complete

* Update instrumentation definitions generator to also implement a `GetIntegrationId()` method

This can be used at runtime by calltarget instrumentation to determine which `IntegrationId` an executing method is associated with. There may be other ways of doing it as well, but given this only needs to be done once per instrumentation type, I think this is sufficient.

Note that the pattern matching code produces lots of duplicates, but the compiler is smart enough to notice they're all the same and merge them into one.

* Exclude ducktype dynamic assemblies from the list of reported dependencies

These add noise, and won't tell us (or customers) much, so ignoring them

* Update CallTarget GetIntegrationOptions implementation

* Bump timeout in TelemetryControllerTests.cs

Saw some strange failures, as well as success at 7.2s, so play it safe

* Don't start telemetry sending background thread until after Tracer.Instance is set

This avoids a double-initialization bug we would run into otherwise

* Update TelemetrySettings to use better URL inferring

If a DD_API key is provided, then assume we're running in an environment without an agent, and send directly to the intake.

If the user provides an override URL, use that.

If not, and the user provides DD_SITE, then use that for the intake. Avoids requiring manual config for telemetry e.g. in AAS (as DD_SITE) must always be configured. If DD_SITE is not provided, fall back to "datadoghq.com"

This logic matches that used in the Rust library.

* Add telemetry key called 'environment_fulltrust_appdomain' to indicate whether the Tracer is operating in an AppDomain with FullTrust (#2339)

* Use ConfigTelemetryData keys instead of strings

* Disable telemetry when using a non-http transport for the agent

* Remove telemetry for SecurityBlockingEnabled

(As this no longer exists)

Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>