Support cleaning filenames in http-client + aspnet
The HttpClient integration (and the ASP.NET/ASP.NET Core integration when the feature flag is disabled) all do "cleaning" of URLs to try and reduce the cardinality introduced by number/guid route parameters. This works by examining each "segment" in a URL, and using heuristics to see if we should replace it with `?`.

This works as expected for the following paths:
- `/some/value/123` => `/some/value/?`
- `/some/value/123/123-456 => `/some/value/?/?`
- `/some/value/123/example.png => `/some/value/?/example.png`

However, it treats the whole final segment as a single unit, which means filenames with extensions will never be obfuscated, e.g.

- `/some-filevalue/123.png`
- `/some-filevalue/123.456`
- `/some-filevalue/123.c`

That can lead to high cardinality, which in turn can cause problems for customers with too many endpoints in their account.

To work around it, we treat the filename as a separate segment from the file extension, and don't try and obfuscate the extension