Add better handling of nullable type chaining (#2424)
* Add better handling of nullable type chaining

Currently, when doing duck chaining, we call `CreateCache<T>.Create(object instance)`
 to do the conversion. Note that this is a boxing operation.
Inside this method, we then call `instance.GetType()` to determine
the type to convert from. However, [doing a boxing operation on a non-null nullable type converts it to the type itself](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/builtin-types/nullable-value-types#how-to-identify-a-nullable-value-type).
The end result is that we create the incorrect proxy mapping.

## An example of the problem

For example, consider this type hierarchy:

```csharp
public class GenericTestClass
{
    public TestStruct? MyTestStruct { get; set; }
}

public struct TestStruct { }
```

Which we want to create a proxy for:

```csharp
[DuckCopy]
public struct GenericProxyStruct
{
    public NullableProxyTestStruct MyTestStruct;
}

[DuckCopy]
public struct NullableProxyTestStruct
{
    public ProxyTestStruct Value;
}

[DuckCopy]
public struct ProxyTestStruct { }
```

Running the following before this PR would throw, as it trys
to proxy `TestStruct` to `NullableProxyTestStruct` _not_ `TestStruct?`

```csharp
var instance = new GenericTestClass { MyTestStruct = new TestStruct() };
var proxy = instance.DuckCast<GenericProxyStruct>();
```

> Note that we currently require an "intermediate" proxy to handle the nullable. It would be nice to handle this implicitly in the future, which incidentally would also solve this issue

The biggest problem is that the behaviour is _different_ depending
on whether the proxied property is `null` at runtime, or it has a value,
so there's no "safe" way to create a proxy heirarchy that works in all cases.

## The fix

The source of this problem is that the `CreateCache<T>.Create(object instance)`
call that is part of `DuckChaining` boxes the source field.
To work around the issue, I created a new, generic, method
`CreateCache<T>.CreateFrom<T>(T instance)` which avoids the boxing. This then creates a proxy between the nullable type as expected,
and the above hierarchy works as expected.

This fix only helps with the duck-chaining case, but this is the
main case in which errors can occur - in most cases, _directly_
duck typing a nullable isn't something we'll need to do (hopefully!)

* No need to box if duck typing a value type with CreateFrom