Update Mongo DB tests to support 2.15.0
Yet another package that doesn't understand SemVer
They deleted a whole bunch of APIs, and some APIs now throw if you use them. Specifically

- KillCursors API has been deleted. I can't find any documentation on how to call this another way.
- $explain modifier isn't supported. I can't find any documentation on how to call this another way.

Given I can't find solution to this in the documentation, I have removed these features when running 2.15.0 and up, and have created a separate snapshot for the latest version. We should reevaluate this later if they improve their docs to try and unify (so we keep testing all the relevant args).

Added a 2.14.x version to the test matrix to make sure we keep testing all code paths