Prepare 2.4.4 hotfix (#2578)
* [Appsec] Update rules file (#2516)

* Update WAF rule-set

Update to the newly released: https://github.com/DataDog/appsec-event-rules/releases/tag/1.2.6

* Improve rule serialization

The WAF uses the same format for rules and data to be scanned. Data to be scanned should be truncated to avoid large requests taking too long to execute. Rules should not be truncated, as this will result in some rules misbehaving.

* Don't trucate strings either

* change rule in unit test, as previous one was removed

* remove unneeded signed / unsigned functions

* updated rule file

* have a different name per system tests artifact if different job attempts

Co-authored-by: Anna <anna.yafi@datadoghq.com>

* [AppSec] Support postfix time format for WAF timeout (#2519)

The system tests set the timeout to '5s', and we don't currently support this format, meaning that we could timeout during the system tests.

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>

* Dont log a warning if no waf timeout has been specified (#2576)

Co-authored-by: Robert Pickering <robert.pickering@datadoghq.com>
Co-authored-by: Anna <anna.yafi@datadoghq.com>
Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>