The goal is to connect the ExplorationTests container (started by docker-compose) to the Azure DevOps dd_agent container.
The problem is that the ExplorationTests container is currently being started by docker-compose in CI, which creates a brand new network, so the traces will not reach dd_agent.

The solution is to try to extend the ExplorationTests with a docker-compose override file to use the existing network instead of the default project network.