Improve source generator verification (#2614)
* Improve source generator verification

Remove the generated files before regenerating them and looking for changes. Previously a change to an input file may mean that a source-generated file _stops_ being generated, and so shouldn't be in the file output, but we would not detect that. By removing the generated files first, we can ensure the output is exactly what we expect.

* Add test file

* Revert "Add test file"

This reverts commit 168a3c58837c9e38935d7ed26e93d72cae758805.