Ensure we test a consistent commit for PRs
By default, Azure Devops clones the special "merge" commit the GitHub creates for PRs. As we clone the repo multiple times (generally, once for each stage), that means later stages can end up cloning a different commit, if there have been commits to master since the first stages run. This often causes failures if those new commits introduce additional tests, or generally change anything.

To ensure we test all stages against the same commit, we make a note of the current commit ID of master before starting the pipeline. If we're running a branch directly, then we use the built-in clone command, so this change is a noop. If we're running a PR however, we clone the PR branch and then merge with the original commit ID _manually_. This ensures we are always testing the same effective merge commit.

The main downside is the additional complexity required in the pipeline, but what's a 100 extra lines in a 2000 line yaml file between friends...