Refactor how we calculate the base endpoint
For some transports (e.g. UDS/named pipes) the "agent URI" is a filepath, not an `http://` path. Currently, many places assume the latter, for example constructing the endpoint by appending a path to it. This is invalid for UDS (e.g. it creates URLs like `C:\temp\path\fdyt3ifhyg.dsf/v4/traces`).

Up to now, this had some strange, but apparently(?) harmless,  effects for UDS, though I'm not entirely convinced on that front.

For example, currently (i.e. on master), we are sending the following headers when making a UDS request on Windows:

   POST C:/v0.4/traces HTTP/1.1
   Host:
   Accept-Encoding:identity
   Content-Length:1
   Datadog-Meta-Lang:.NET
   Datadog-Meta-Tracer-Version:2.7.0.0
   x-datadog-tracing-enabled:false
   Datadog-Meta-Lang-Interpreter:.NET
   Datadog-Meta-Lang-Version:6.0.4
   Datadog-Client-Computed-Top-Level:1
   Content-Type:application/msgpack

Notice the query path and Host tags are wrong...

With these changes, we use localhost as the Hostname for requests where the "host" doesn't make sense (named pipes and uds). This also resolves the "path" issue.