Activate and expose the WAF's obfuscator
The WAF now comes with a build in obfuscator based on RE2 patterns to
scan the output of the WAF for possible sensitive.

This PR activates that functionality by setting the patterns to the
standard values.

It also allows users to control the obfuscation via two environment
variables: `DD_APPSEC_OBFUSCATION_PARAMETER_KEY_REGEXP` and
`DD_APPSEC_OBFUSCATION_PARAMETER_VALUE_REGEXP` (and by user all the
other standard ways of setting config in .NET)