Write log files directly to build-data, instead of system
There's something weird going on with the hosted agents, where we're seeing logs from other pipeline runs. To try and avoid that, set `DD_TRACE_LOG_DIRECTORY` so logs are written directly to the build_data folder, instead of copying them later. That should avoid the issue (whatever it is)