Build in support for telemetry testing into `MockTracerAgent`
When we're sending to the "agent" endpoint, we need to reuse the same Socket/HttpListener for both traces and telemetry, so we need to incorporate the `MockTelemetryAgent` functionality into the `MockTracerAgent`, and split based on path.

This required adding support for parsing the URL requested from the HTTP request in `MockHttpParser`.