Improve git checkout to handle rebases on long-lived clones
The benchmarks etc VMs are long-lived, so we re-clone the git repo.
However, this can cause issues when references change (e.g. rebasing/merges) and can cause the merge to fail.

The "fix" in this commit is the addition of the `git reset origin/$prBranch --hard`. That ensures we are correctly mirroring the remote before we do the merge