Fix ThreadAbortAnalyzer
With the latest .NET SDK (6.0.300) we were getting the following inexplicable message:

```
error AD0001: Analyzer 'Microsoft.CodeAnalysis.Analyzers.MetaAnalyzers.DiagnosticDescriptorCreationAnalyzer' threw an exception of type 'System.ArgumentException' with message 'Syntax node is not within syntax tree'.
```

After some trial and error, I narrowed this down to us storing some strings as constants (I know, it's terrible).

Also decided to disable the "AnalyzerReleases" analyzer as it adds nothing but noise.

With this change we should be able to update to the latest .NET SDK