Fix incorrect `http.status_code` in classic mode
In ASP.NET MVC, when an exception is  thrown and IIS is running in "classic" mode, the status code is not set on the `HttpContext.Response.StatusCode` property. That means we default to a 200 response code when the request ends, even though IIS will return a `500` response.

This PR checks for this eventuality both in the Module and in the MVC integration.

Note that the WebApi2 integration snapshots are _already_ correct, so the behaviour is clearly different in that case.