Fix flaky LambdaRequestBuilderTests
We have been seeing flaky behaviour from the recent merge of `LamdbdaRequestBuilderTests`, failing with the following

```
System.NotImplementedException : This property is not implemented by this class.
  at System.Net.WebRequest.set_Method(String value)
  at Datadog.Trace.ClrProfiler.ServerlessInstrumentation.AWS.LambdaRequestBuilder.Datadog.Trace.ClrProfiler.ServerlessInstrumentation.AWS.ILambdaExtensionRequest.GetEndInvocationRequest(Scope scope, Boolean isError)
```

This is quite a strange error, as it implies we're running with a `WebRequest` _not_ an `HttpWebRequest` as you'd expect.

My assumption is that this is a parallelization issue with the `ApiWebRequestFactoryTests` which overrides the `http://` prefix for WebRequest. If any test that uses WebRequest is running at the same time at ApiWebRequestFactoryTests we will get flaky behaviour

To work around it, added a [WebRequestCollection] collection, and disabled parallelization.

If you can avoid using WebRequest, please do :wink: