Don't run all the installer smoke tests on all the PRs
With all the extra jobs there's more chance of random flakiness due to connectivity issues etc. In addition we are generally unlikely to make changes that _only_ affect the installers, and don't break other things.

To reduce the impact of flake, we skip most of the installer tests for PRs, and instead run them on master only. You can also trigger a "full run" by setting the variable `isRunAllInstallerTestsBuild` in AzDo.

I opted to _keep_ running the "raw" installer tests on arm64 + linux, and the "tracer home" tests on Windows, so that we are at least doing some installer testing on every PR