Add support for incrementally generating a hash
This allows you to pass the "previous" hash version to the GenerateHash version. This will be useful as we need to concatenate multiple strings together to create the final hash value. By using these "incremental" versions, we can generate the hash in a zero-allocation way (for small enough strings in netcoreapp3.1+).