Installer test tweaks (#2963)
* Reduce the timeouts on the installer tests

Azure Devops seems especially prone to losing contact with agents when running these tests. In that situation, the job will continue until the timeout expires. That can mean a long time before you get feedback. Reducing the timeout should reduce the feedback loop. The rough expected times + timeout set are:
- Linux max ~5 mins (10 min timeout)
- Arm64 max ~15 mins (30 min timeout)
- Windows max ~15 mins (30 min timeout)

* Don't run all the installer smoke tests on all the PRs

With all the extra jobs there's more chance of random flakiness due to connectivity issues etc. In addition we are generally unlikely to make changes that _only_ affect the installers, and don't break other things.

To reduce the impact of flake, we skip most of the installer tests for PRs, and instead run them on master only. You can also trigger a "full run" by setting the variable `isRunAllInstallerTestsBuild` in AzDo.

I opted to _keep_ running the "raw" installer tests on arm64 + linux, and the "tracer home" tests on Windows, so that we are at least doing some installer testing on every PR