Ignore process_id in snapshots on Linux
When running in Docker, the pid of the main process is always 1.0, however, when we use `dd-trace` to instrument an app, the pid is sometimes _not_ 1.0.

I'm not entirely sure why it's _sometimes_ 1.0, and sometimes not, but I would _guess_ that it's some sort of race. Either way, I think it's safe to ignore it in the snapshots.