[Tracer] Add Fluent API to validate integration spans and produce documentation of tracing integrations -- Redo (#3011)
I created C# APIs to easily and clearly chain together assertions for span properties and tag values. By limiting the API surface area, I was also able to write a build task to parse this file and generate a markdown file so we can have an easily linked file in our repo to show what tracing information we generate.

All of the integrations to date are being tested with this readable testing framework.