Always use native loader for tests (part of package layout update) (#3036)
* Remove USE_NATIVE_LOADER from test code

We should always be testing against the native loader

* Replace usages of GetTracerNativeDLLPath with GetNativeLoaderPath() in tests

Ideally, we don't want to hard-code the paths for both the native loader and the native tracer. The loader.conf file contains relative paths, so we should rely on the values embedded in there

There _are_ still some hardcoded paths due to the DD_DOTNET_TRACER_HOME directory

* Update arm64 throughput tests to use the native loader

* Stop MultipartFormTests from inheriting from TestHelper

TestHelper is only for integration tests, as it relies on the tracer/native loader etc having been built.