Centralise handling of HTTP requests in MockTracerAgent
We have multiple implementations of the MockTracerAgent to handle each of the different transports
- Tcp/udp
- Named pipes
- UDS

The Named Pipes and UDS implementations were sharing some of their "http request deserialization" code, but the TCP transport was completely separate. More over, the TCP transport supported many extra features: config, debugger, stats etc.

This PR centralises the parsing of the HTTP requests, so the logic can be reused across implementations.