Capture and tag the Kafka consumer group on comsumer spans
We need to capture the consumer group for data streams monitoring

This is annoyingly hard, as it's only exposed indirectly in the `Consumer` constructor, in the config provided by `ConsumerBuilder`, and is not accessible later in the `Consume` method.

To work around that, using a `ConditionalWeakTable` to create a map from `Consumer` instance to a group ID. We then look up the group ID again when `Consume` is called. We explicitly remove the entry from the table when the consumer is closed or disposed.