Add support for Data Streams Monitoring to Kafka
If enabled, checkpoints and injects the pathway context into Kafka headers when producing Kafka messages.
If enabled, extracts, merges (for fan-in situations), and sets a checkpoint when consuming.

As far as I can tell, in the "default" scenarios (`KafkaCreateConsumerScopeEnabled` is true and processing in a loop; `KafkaCreateConsumerScopeEnabled` is false and consuming multiple before processing) we will never need to "merge" contexts. Instead, we may have to expose the "merge" functions for users running a fan-in scenario. Need to think about the API to expose for that, and somehow incorporate into the ParentContext extraction helper