Refactor the exclude list for assemblies
In https://github.com/DataDog/dd-trace-dotnet/pull/1663 we switched from excluding the Microsoft.Extensions prefix to excluding each of the non-logging assemblies individually.

This is not future proof (as new libraries could be added), and increases the number of assemblies we need to match against.

As a workaround, list adds an explicit "include" list for assemblies which match the "exclude" prefix, but which should be included anyway.