[ASM] 1-Click activation (#3081)
## Summary of changes
Adds the ability to enable and disable the ASM on the fly

## Reason for change
Remote configuration will allow to fine tune the behavior of the agent, including the activation or deactivation of security inside the application.

## Implementation details
`Security` class implements `UpdateStatus` based on the `settings.Enabled` flag which registers or unregisters the waf triggers in the request. Furthermore and in order to be as efficient as possible, the Profiler is told to instrument or deinstrument the ASM owned CallTargets. This forces a Rejit or a Restore of the designated CallTargets in runtime. To allow deinstrumentation of CallTargets Profiler has implemented a new interop function called `RemoveCallTargetDefinitions`

## Test coverage
UnitTest for activation and deactivation of CallTargets in `CallTargetNativeTest` 
AspNetCore5 integration tests added in `Samples.Security.AspNetCore5`

## Other details
Fixes a problem found when trying to reuse the auto generated CallTargetDefinitions, which were disposed right after their first use, leaving them in the catalog dictionary in zombie status.