Implement rcte2 - Configuration apply status (#3175)
## Summary of changes

A system that allows a remote config client to acknowledge each file provided by RC.

## Reason for change

It will allows us to support customers using remote config more easily.

## Implementation details

As remote config is event based, acknowledging a file relies on the subscriber of the products event to call a method on the event args class to acknowledge successful application of the config. If an error was thrown then the config will automatically be set into an error state. A subscribe can also explicitly set an error if they whish.

## Test coverage

The feature is covered by an integration test, which includes a new mechanism for recovering the remote config request.

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>