Update StatsTests so it correctly shuts down all background processes
The StatsTests create standalone (i.e. non-static-global) `Tracer` instances, which in turn creates `TracerManager` instances. The tests were specifically flushing and closing the associated AgentWriter (by way of a helper internal method on `Tracer`), but this left all the other background services (e.g. the `DiscoveryService`) running in the background. When the test finished, and the agent shutdown, those services would keep running in the background, trying to hit a non-existent agent, and writing a log each time