Remote Configuration Tracer Extension 3 - Remote Configuration Client capabilities (#3186)
## Summary of changes

RCTE3 allows a library to advertise what remote config features (capabilities) that it supports. 

## Reason for change

It will be helpful for clients to be able to see in the interface if they can make a change to a services via remote config or not.

## Implementation details

The implementation is mostly straight forward. The only complicated part is a reference to the `RemoteConfigurationManager` class is required at startup and the initialization order doesn't seem to be deterministic. To get round this I wrote a small mechanism that checks if `RemoteConfigurationManager` is initialized and queues the request for a reference if it isn't. We may need a generalization similar to this, for other services, in the future.

## Test coverage

Tests piggy back on the "toggle" tests for activating and deactivating ASM via remote config. This allows us to check the capabilities are correctly sent with each request.

Co-authored-by: Anna <anna.yafi@datadoghq.com>