IP Blocking for .NET Framework (#3207)
## Summary of changes

This builds on https://github.com/DataDog/dd-trace-dotnet/pull/3066 to bring blocking capabilities to the .NET Framework as well as core. 

This provides the ability to block a request but is not linked to any real mechanism to block requests, only test blocking rules exist for now.

## Reason for change

We want to support blocking on both .NET Framework and Core.

## Implementation details

Most of the blocking mechanism was put in place in https://github.com/DataDog/dd-trace-dotnet/pull/3066, this just fills in the missing parts for .NET Framework.

## Test coverage

Covered by integration tests.