[Tracer] Add support for instrumenting interface methods (#3256)
Key changes
- Adds new category of definitions using new enum value ClrProfiler.IntegrationType.Interface
- Interface integrations are registered to the CLR profiler with a new PInvoke
- When processing a module for interface instrumentation, re-use part of the derived instrumentation to efficiently determine if an assembly reference to the interface assembly exists and instrument methods that implement an interface method
- Update the RabbitMQ implementation and test cases to use interface instrumentation, allowing applications to provide their own types to implement the RabbitMQ message consumer interface

Caveats
One case will not be instrumented through interface instrumentation, and instead requires derived instrumentation: When a base type implements an interface, and a derived type overrides the implementing method. In this case, we cannot use the metadata at module load to know that this particular method is overriding an interface method. To instrument cases like these, we need to add a derived instrumentation that targets the specific base type.

Additional changes
- Refactors the integration definition source generator so it easier to update. This results in some additional whitespace in the generated file but that seems like a worthy tradeoff
- Updates the RabbitMQ.DistributedTracing samples